refactor: merge input components
